📝 :: View 수정
